+++ vrp-pragmatic/src/constraints/breaks.rs
@@ -3,7 +3,7 @@
 mod breaks_test;
 
 use crate::constraints::*;
-use std::collections::HashSet;
+use hashbrown::HashSet;
 use std::slice::Iter;
 use std::sync::Arc;
 use vrp_core::construction::constraints::*;