+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - `breaking`: job skills are now defined by `allOf`, `oneOf` and `noneOf` conditions.
+- use `hashbrown` library in `pragmatic` crate
 
 ## [v1.6.1] - 2020-10-15